% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_package.R
\name{check_package}
\alias{check_package}
\title{Make sure a package is available}
\usage{
check_package(
  pkg_name,
  msg = "",
  install_directions = glue::glue("install.packages('{pkg_name}')")
)
}
\arguments{
\item{pkg_name}{the package to check}

\item{msg}{a helpful message preappended to the error to explain why package
is needed}

\item{install_directions}{code to install the packages - defaults to
`install.packages`}
}
\value{
invisibly returns TRUE if package is available otherwise throws an
  error
}
\description{
This function checks to see if a package is installed an throws an error if
it is not.
}
\examples{
check_package('base')
}
