% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkMinRate.R
\name{checkMinRate}
\alias{checkMinRate}
\title{Include only the risk factors that appear in x\% of the data.}
\usage{
checkMinRate(x, min_rate = 0.01)
}
\arguments{
\item{x}{The object to be evaluated.}

\item{min_rate}{The threshold for permissable risk factor found in patient population. Defaults to 1\%.}
}
\value{
Returns variables that meet the min_rate criteria.
}
\description{
\code{checkMinRate} returns a logical value after evaluating an object whose existence is greater than a specified amount.
Despite permitting any object class to be passed, only an object
with a class of 'logical'will be evaluated.
}
\examples{
data <- data.frame(diab = c(T, T, F, F), obese = c(T, F, T, F), site = c("A", "B", "C", "D"),
     bmi = c(16, 18, 20, 22))
checkMinRate(data$diab, min_rate = .45)
}
