% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fillDown.R
\name{fillDown}
\alias{fillDown}
\title{fillDown fills a measurement down for a specific amount of time.}
\usage{
fillDown(meas_col, time_col, carry_forward = lubridate::hours(6))
}
\arguments{
\item{meas_col}{column to carry forward}

\item{time_col}{column that defines what "time" a measurement is taken. Can be
any data type that supports numeric operations. This column must be ordered ascending.}

\item{carry_forward}{how long to carry a measurement forward for. Must be the
same type as `time_col`}
}
\value{
vector type same as meas_col
}
\description{
For example, if you have a lab measurement and want it to be valid for the
next 6 hours this function can be used to set the next 6 hours if no other
measurement has been made. Only NA's are filled in.
}
\examples{
library(dplyr)
set.seed(2)
d <- data.frame(meas = rnorm(100),
                time = 1:100)
d$meas[sample(1:nrow(d), 70)] <- NA
d \%>\%
  mutate(filled_meas = fillDown(meas, time, carry_forward = 3))
}
