% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_tools.R
\name{lkp_to_factor}
\alias{lkp_to_factor}
\title{Relabel factors using a named vector}
\usage{
lkp_to_factor(x, lkp_vals, ..., check_names = TRUE)
}
\arguments{
\item{x}{a character or factor vector to modify}

\item{lkp_vals}{either a named vector where the names are the values in x and
the values are the new levels or a character vector defining the order of
the values of x.}

\item{...}{additional parameters passed to \code{factor()}}

\item{check_names}{check that all the names exist}
}
\value{
a factor
}
\description{
Relabel factors using a named vector
}
\examples{
x <- c("apple", "bear", "banana", "dear")
# new value is "fruit"
levels <- c("apple" = "fruit", "banana" = "fruit")
lkp_to_factor(x, levels)

# just reorder factors
lkp_to_factor(x, c("dear", "banana", "apple", "bear"))
}
