% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typeConversions.R
\name{log2fact}
\alias{log2fact}
\title{log2fact converts all logical columns in a dataframe to a 2-level factor.}
\usage{
log2fact(df, positive = "Yes", negative = "No")
}
\arguments{
\item{df}{a dataframe where you want to convert}
}
\value{
the full df with factor columns instead of logical.
}
\description{
This is useufl for on-the-fly conversion for data profiling
}
\examples{
d <- data.frame(a = rnorm(100),
                b = sample(100, c(T, F)))
log2fact(d)

}
