% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maskData.R
\name{maskData}
\alias{maskData}
\title{maskData obfuscates a column vector}
\usage{
maskData(x, header = "", randomize = TRUE)
}
\arguments{
\item{x}{column vector of data to mask}

\item{header}{character string to preappend to each value}

\item{randomize}{should the mask be applied randomly?}
}
\value{
a named character vector
}
\description{
maskData maps the current values in a column (such as names) and maps them to
either 1 or 2 levels. This is useful for quickly changing surgeon names in a
report so that the surgeon identity is protected. The orignal values are set
to the names of the newly created character vector. This function can map up
to 26^2 (676) values.
}
\examples{
test_df <- data.frame(original = sample(paste0("SURG",1:total_surgeons),30,
replace = TRUE),
stringsAsFactors = FALSE)

test_df <- dplyr::mutate(test_df, new = maskData(original))
test_df
}
