% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dateOps.R
\name{personAge}
\alias{personAge}
\title{Calculate age based on start and stop dates}
\usage{
personAge(from, to)
}
\arguments{
\item{from}{birthdate}

\item{to}{given date}
}
\value{
numeric
}
\description{
Calculate age based on start and stop dates
}
\examples{
# stay tuned

}
