% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qccOperations.R
\name{plot_gg}
\alias{plot_gg}
\alias{plot_gg.qcc}
\title{plot qcc using ggplot}
\usage{
plot_gg(x, ...)

\method{plot_gg}{qcc}(
  x,
  add.stats = TRUE,
  chart.all = TRUE,
  label.limits = c("LCL ", "UCL"),
  gr_func = as.numeric,
  ...
)
}
\arguments{
\item{x}{a qcc object}

\item{...}{currently not used}

\item{add.stats}{currently not used}

\item{chart.all}{currently not used}

\item{label.limits}{a character vector specifying the labels for control limits.}

\item{gr_func}{function to apply to group labels, useful for dates or
numbers. Set this to as.numeric or as.Date when your groups are
numbers/dates respectively}
}
\value{
a ggplot2 plot object
}
\description{
plot qcc using ggplot
}
\section{Methods (by class)}{
\itemize{
\item \code{qcc}: for qcc functions
}}

\examples{

data("pistonrings", package = 'qcc')
diameter <- qcc::qcc.groups(pistonrings$diameter, pistonrings$sample)
qcc_obj <- qcc::qcc(diameter[1:25,], type="xbar")
plot_gg.qcc(qcc_obj)

qcc_new <-  qcc::qcc(diameter[1:25,], newdata = diameter[24:40,], type="xbar")
plot_gg.qcc(qcc_new)
}
