% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dateOps.R
\name{pretty_date}
\alias{pretty_date}
\title{Create a full date with an ordinal date}
\usage{
pretty_date(x = Sys.Date(), month_code = "\%B", year_code = "\%Y")
}
\arguments{
\item{x}{a date object to convert to a string}

\item{month_code}{the code used to format the month - defaults to full month
name}

\item{year_code}{the code used to format the year - defaults to full 4 digit
year}
}
\value{
a length 1 character vector
}
\description{
This function is useful for formatting dates in reports - especially in the
header of an Rmarkdown report. Default behavior uses today's date. See
\code{\link{strptime}} for details on month and year code strings.
}
\examples{
# returns nice format for today's date
pretty_date()

pretty_date(as.Date('2020-03-01'))
# March 1st, 2020
}
