% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probodds.R
\name{prob2odds}
\alias{prob2odds}
\alias{prob2logodds}
\alias{odds2prob}
\alias{logodds2prob}
\title{Helper functions to convert probabilities and odds}
\usage{
prob2odds(p)

prob2logodds(p)

odds2prob(x)

logodds2prob(lx)
}
\arguments{
\item{p}{a probability}

\item{x}{an odds}

\item{lx}{a log-odds}
}
\value{
a numeric vector
}
\description{
Helper functions to convert probabilities and odds
}
\section{Functions}{
\itemize{
\item \code{prob2logodds}: convert to logodds

\item \code{odds2prob}: convert odds to probability

\item \code{logodds2prob}: convert logodds to probability
}}

\examples{
prob2odds(.5)
odds2prob(prob2odds(.5))
}
