% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qccOperations.R
\name{qcc.formula}
\alias{qcc.formula}
\title{Formula interface to qcc}
\usage{
qcc.formula(formula, data, type, plot = FALSE, cutoff = NULL, ...)
}
\arguments{
\item{formula}{a formula describing the outcome variable and the grouping
variable}

\item{data}{a data frame}

\item{type}{a character string specifying the group statistics to compute. \cr
  Available methods are:

\tabular{lll}{
              \tab Statistic charted   \tab Chart description \cr
\code{"xbar"} \tab mean                \tab means of a continuous process variable \cr
\code{"R"}    \tab range               \tab ranges of a continuous process variable \cr
\code{"S"}    \tab standard deviation  \tab standard deviations of a continuous variable \cr
\code{"xbar.one"} \tab mean            \tab one-at-time data of a continuous process variable \cr
\code{"p"}    \tab proportion          \tab proportion of nonconforming units \cr
\code{"np"}   \tab count               \tab number of nonconforming units \cr
\code{"c"}    \tab count               \tab nonconformities per unit \cr
\code{"u"}    \tab count               \tab average nonconformities per unit \cr
\code{"g"}    \tab count               \tab number of non-events between events
}

Furthermore, a user specified type of chart, say \code{"newchart"}, can be provided. This requires the definition of \code{"stats.newchart"}, \code{"sd.newchart"}, and \code{"limits.newchart"}. As an example, see \code{\link[qcc]{stats.xbar}}.
}

\item{plot}{logical. If \code{TRUE} a Shewhart chart is plotted.}

\item{cutoff}{a cutoff date for new data}

\item{...}{further arguments passed to \code{qcc:qcc()}}
}
\value{
a qcc object
}
\description{
This function provides a simplified interface to the qcc function. The main
goal is to remove the need for using the qcc groups function. This allows for
much simpler use of the function within piping functions/map.
}
\examples{
data("pistonrings", package = 'qcc')
diameter <- qcc::qcc.groups(pistonrings$diameter, pistonrings$sample)

# Without Cutoff
qcc_obj <- qcc::qcc(diameter[1:25,], type="xbar")
qcc_obj2 <- qcc.formula(diameter ~ sample, data = dplyr::filter(pistonrings, sample <= 25), type = "xbar", plot = TRUE)

# With Cutoff
qcc_new <-  qcc::qcc(diameter[1:25,], newdata = diameter[24:40,], type="xbar")
qcc_new2 <- qcc.formula(diameter ~ sample, data = pistonrings, type = "xbar", plot = TRUE, cutoff = 25)

# Dates and p chart
## use POSIXct, not Date object to create date
start <- as.POSIXct('2017-01-01')
end <- as.POSIXct('2017-01-15')
samp_dates <- seq(from = start, to = end, length.out = 15)
samp_dates <- sample(samp_dates, 100, replace = TRUE)
old_process <- samp_dates < '2017-01-08'
vals <- ifelse(old_process,rbinom(sum(old_process), 1 , p = 0.5), rbinom(sum(!old_process), 1,  p = 0.2) )
d <- data.frame(samp_dates, vals)

## Run the graphs with plot = TRUE
qcc.formula(vals ~ samp_dates, data = d, type = "p", plot = TRUE)
qcc.formula(vals ~ samp_dates, data = d, type = "p", plot = TRUE, cutoff = '2017-01-08')
}
