% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantileCut.R
\name{quantileCut}
\alias{quantileCut}
\title{quantileCut cuts a numeric vector into the a prespecified number of breaks.}
\usage{
quantileCut(
  x,
  number_breaks = 10,
  force_unique = TRUE,
  silent = TRUE,
  na.rm = FALSE,
  ...
)
}
\arguments{
\item{x}{a numeric vector}

\item{number_breaks}{how many groups should there be}

\item{force_unique}{should you make sure the breaks are unique?}

\item{silent}{will not warn if breaks aren't unique}

\item{na.rm}{will na's be removed?}

\item{...}{passed to \code{\link{cut}}}
}
\value{
a factor (result of \code{cut})
}
\description{
This function is useful for breaking up data into equal unique parts
}
\examples{
library(dplyr)
tibble(original = rnorm(999)) \%>\%
  mutate(grouped = quantileCut(original)) \%>\%
  group_by(grouped) \%>\%
  summarise(MEAN = mean(original),
            MEDIAN = median(original),
            SD = sd(original),
            CT = n())
}
