% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replaceVal.R
\name{replaceValue}
\alias{replaceValue}
\title{Simple `ifelse`}
\usage{
replaceValue(check, replace, replace_if = NA)
}
\arguments{
\item{check}{vector to check equality to `replace_if`}

\item{replace}{replacement for `check`. Recycled if necessary}

\item{replace_if}{currently only allows a length = 1 vector to compare to `check`}
}
\value{
vector of same type as `check`
}
\description{
replaceVal acts very similary to `ifelse` but with a more compact syntax for checking a column against a single value.
}
\examples{
\dontrun{
}
}
