% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runLocalScript.R
\name{runLocalScript}
\alias{runLocalScript}
\title{runLocalScript returns a tidy output for a file that was run}
\usage{
runLocalScript(
  x,
  args = "--vanilla",
  std_err = NULL,
  std_out = NULL,
  error_on_missing = FALSE
)
}
\arguments{
\item{x}{a path to a file to run}

\item{args}{command line arguments for R}

\item{std_err}{either NULL (writes to temp file) or values allowed by \link[sys]{r_wait}}

\item{std_out}{either NULL (writes to temp file) or values allowed by \link[sys]{r_wait}}

\item{error_on_missing}{if FALSE then the function catches errors and returns
a reasonable tibble otherwise an error is thrown.}
}
\value{
a tibble with the following columns:
    status - the exit status of the script
    std_error - the output of std_error
    std_out - the output of std_out
    time_taken - the amount of time it took to run the function
}
\description{
This function returns a tibble with all of the information available from
running the file. Specifically it helps capture the standard error/output by
dumping to a temporary file and reading it back in. It's also possible to
specify the std_out and std_error files if desired.
}
