% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setOptions.R
\name{theme_vakdr}
\alias{theme_vakdr}
\title{Default Options for automatic loading of vakdr}
\usage{
theme_vakdr(...)
}
\arguments{
\item{...}{passed to theme_minimal}
}
\value{
a ggplot2 theme object
}
\description{
The functions below represent default options for graphing/knitting that
should be used when creating any report.
}
\details{
\describe{

\item{`theme_vakdr`}{ A modified version of \code{ggplot2::theme_minimal()}.
See source code of theme_vakdr for more information, and \code{\link[ggplot2]{theme}}
for details.}

}
}
\examples{
library(ggplot2)
ggplot(mtcars, aes(x = hp, y = mpg, color = as.factor(cyl))) + geom_point()
# Revert back to default ggplot2 theme:
theme_set(theme_gray())
ggplot(mtcars, aes(x = hp, y = mpg, color = as.factor(cyl))) + geom_point()

}
