% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toClipboard.R
\name{toClipboard}
\alias{toClipboard}
\title{Copy a data.frame to clipboard for pasting into Excel}
\usage{
toClipboard(
  x,
  file = "clipboard",
  sep = "\\t",
  row.names = FALSE,
  na = "",
  ...
)
}
\arguments{
\item{x}{the object to be written, preferably a matrix or data frame.
    If not, it is attempted to coerce \code{x} to a data frame.}

\item{file}{either a character string naming a file or a \link{connection}
    open for writing.  \code{""} indicates output to the console.}

\item{sep}{the field separator string.  Values within each row of
    \code{x} are separated by this string.}

\item{row.names}{either a logical value indicating whether the row
    names of \code{x} are to be written along with \code{x}, or a
    character vector of row names to be written.}

\item{na}{the string to use for missing values in the data.}

\item{...}{other parameters passed to write.table}
}
\description{
This function wraps around \code{write.table} to push a data frame to
clipboard. It is pipable. Please note that large data.frames will cause the
clipboard to crash.
}
\examples{
toClipboard(mtcars)
}
