% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{windowThreshold}
\alias{windowThreshold}
\title{windowThreshold Function}
\usage{
windowThreshold(meas_col, meas_norm, time_col, look_backward, look_forward)
}
\arguments{
\item{meas_col}{the numeric column of the measure}

\item{meas_norm}{a numeric vector showing the c(low, high) acceptable range}

\item{time_col}{the time column, as a numeric}

\item{look_backward}{the amount to look backward for an invalid value}

\item{look_forward}{the amount to look forward for an invalid value}
}
\value{
logical vector for each row in the initial time row
}
\description{
windowThreshold Function
}
\examples{
d <- data.frame(
measure = 1:10,
time = 0:9
)
library(dplyr)
d \%>\% mutate(hit = windowThreshold(measure, c(2,8), time, 2, 0))


}
