% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeXLSX.R
\name{writeXLSX}
\alias{writeXLSX}
\title{writeXLSX is a convenient way to write dataframes to xlsx files.}
\usage{
writeXLSX(
  x,
  file = getOption("vakdr.writeXLSX"),
  sheetName = deparse(substitute(x)),
  col.names = TRUE,
  row.names = FALSE,
  showNA = FALSE,
  addFilter = TRUE,
  autoSize = TRUE
)
}
\arguments{
\item{x}{a \code{data.frame}.}

\item{file}{the file name to write to}

\item{sheetName}{the sheet name to write the data frame to}

\item{col.names}{a logical value indicating if the column names of
    \code{x} are to be written along with \code{x} to the file.}

\item{row.names}{a logical value indicating whether the row names of 
    \code{x} are to be written along with \code{x} to the file.}

\item{showNA}{a boolean value to control how NA's are displayed on the
    sheet.  If \code{FALSE}, NA values will be represented as blank
    cells.}

\item{addFilter}{should an auto filter be added to each column heading?}

\item{autoSize}{should each column be autosized?}
}
\description{
The default options write a data frame to sheet named after the object name
in a file `data_output.xlsx`. You can write many sheets to the same file by
simply passing either differently named objects or passing a different name
for the sheet. Note that sheets will be overwritten without warning. New
sheet wil always be added to a file (old sheets are not cleared out).
}
\examples{
\dontrun{
writeXLSX(mtcars)
}
}
