% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dateOps.R
\name{personAge}
\alias{personAge}
\title{Calculate age based on start and stop dates}
\usage{
personAge(from, to)
}
\arguments{
\item{from}{birthdate}

\item{to}{given date}
}
\value{
numeric
}
\description{
This function makes sure that a person only becomes older on their birthdate.
}
\examples{
birth_date <- as.Date('1978-02-22')
personAge(birth_date, '2021-02-22') #43

}
