% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyrExtend.R
\name{rollup}
\alias{rollup}
\title{rollup add a total row to a dataframe}
\usage{
rollup(.data, ..., label = "Total")
}
\arguments{
\item{.data}{A data frame, data frame extension (e.g. a tibble), or a
lazy data frame (e.g. from dbplyr or dtplyr). See \emph{Methods}, below, for
more details.}

\item{...}{<\code{\link[dplyr:dplyr_data_masking]{data-masking}}> Name-value pairs of summary
functions. The name will be the name of the variable in the result.

The value can be:
\itemize{
\item A vector of length 1, e.g. \code{min(x)}, \code{n()}, or \code{sum(is.na(y))}.
\item A vector of length \code{n}, e.g. \code{quantile()}.
\item A data frame, to add multiple columns from a single expression.
}}

\item{label}{what should the total row be labeled as? This is applied to all
grouping variables that are characters or factors.}
}
\value{
An object of the same class as .data. One grouping level will be
  dropped.
}
\description{
This function is inteded to extend the `summarise` function in dplyr by
binding (using `bind_row`) a "Total" row to the dataframe. This is especially
useful when generating reports that require a summary line. It also
appropriately scales everything to make sure that aggregate functions such as
`mean` are taken on the raw data and not the aggregate data.
}
\examples{
library(dplyr)

# cyl is a number so it's not possible to label the "Total Row"
mtcars \%>\%
  group_by(cyl) \%>\%
  rollup(mean = mean(disp), n = n())

# Casting cyl to a character allows automatic labeling of the final row.
mtcars \%>\%
  mutate(cyl = as.character(cyl)) \%>\%
  group_by(cyl) \%>\%
  rollup(sum = sum(disp),
         mean = mean(disp),
         sd = sd(disp),
         n = n())

# Set label = NA to not label the last row (same as if it's a number)
mtcars \%>\%
  mutate(cyl = as.character(cyl)) \%>\%
  group_by(cyl) \%>\%
  rollup(sum = sum(disp),
         mean = mean(disp),
         sd = sd(disp),
         n = n(),
         label = NA)
}
